Only delete the idigi process if it is still running.

+++ test.py
@@ -77,8 +77,10 @@ def run_tests():
             print '>>>Executing [%s]' % test_script
             rc = os.system('export PYTHONPATH=../;cd %s; nosetests --with-xunit %s' % (TEST_SCRIPT_DIR, test_script))
 
-        p = int(pid)
-        os.kill(p, signal.SIGKILL)
+        if pid != '':
+            p = int(pid)
+            os.kill(p, signal.SIGKILL)
+
         os.system('cd ../../../../../')
 
         if rc != 0:
